refactor to support multiple binaries sharing common internal packages (#15)
