+++ .gitignore
@@ -1,1 +1,2 @@
 *.env
+**/.DS_Store
